Major refactoring of page navigation in the admin.

+++ oy/contrib/collect/storages/local.py
@@ -9,7 +9,6 @@
     :license: MIT, see LICENSE for more details.
 """
 
-import os
 import shutil
 from pathlib import Path
 from .base import BaseStorage
@@ -17,10 +16,18 @@ from .base import BaseStorage
 
 class LocalFileSystemStorage(BaseStorage):
 
+    name = "Local File System Storage"
+
     def run(self):
         for src, dst in self:
             src, dst = Path(src), Path(dst)
+            if dst.exists() and (src.stat().st_mtime < dst.stat().st_mtime):
+                self.log(
+                    f"Ignoring file {src}. The recent version of the file already exists in the destination directory."
+                )
+                continue
             parent = dst.parent
             if not parent.is_dir():
                 parent.mkdir(parents=True, exist_ok=True)
             shutil.copy(src, dst, follow_symlinks=False)
+            self.log(f"Copied {src} to {dst}.")
